Resolved not being able to find ANY templates.
Turns out, what was causing it was the "load task macros" command I hadnt yet deleted on the dashboard html. When I deleted the task macros file I expected it to throw out an error whenever it couldnt locate the task macros. It was throwing those exceptions, but this was then being nested and obscured by djangos system of exception catching. I created a util to help debug the problem and eventually got to the bottom of it.