Use filter instead of exclude for our homepage
It's been a long time since I touched this code, so it's hard for me
to understand exactly whyyyyy stuff is. I can only really discern
what's gone wrong.

Case-in-point - the code for the home page was throwing up an
exception after a Django update. The code leading up to that is a
complicated bit, but it seems like I was ordering stuff very
specifically using a UNION.

However, after a Django update, subsequent exclusion calls were
creating exceptions. I think this can be patched using filter.

I say think because, honestly, I have no way to test this other than
pushing the code to a staging area and testing. My knowledge of this
program is and Django is all so degraded. I can't test locally, my
development environment is entirely degraded. I just hope this works.